requirements.txt
docker-py
redis